fix(calendar): fix after merge conflict
