+++ packages/calendar/src/components/calendar-mobile/index.ts
@@ -1,2 +1,3 @@
 export * from './Component';
 export * from './typings';
+export * from './components/calendar-month-only-view';
